requirements.txt
Pillow ~= 8.0
numpy ~= 1.19